LayoutTests/imported/w3c/web-platform-tests/css/css-color-hdr/inheritance-expected.txt


FAIL Property dynamic-range-limit has initial value no-limit assert_equals: expected "no-limit" but got "constrained-high"
PASS Property dynamic-range-limit inherits
